chore: convert setup-test files to js
