+++ .eslintrc.cjs
@@ -26,7 +26,7 @@ module.exports = {
         chai: true,
     },
     parserOptions: {
-        ecmaVersion: 2018,
+        ecmaVersion: 2020,
     },
     rules: {
         indent: ['error', 4, {
