+++ .mocharc.integration.json
@@ -2,7 +2,7 @@
     "extension": ["js", "ts"],
     "spec": "packages/**/integration.test.ts",
     "require": "ts-node/register",
-    "file": ["test--registers.integration.ts"],
+    "file": ["test--registers.integration.js"],
     "node-option": [
         "experimental-specifier-resolution=node",
         "loader=ts-node/esm"
